ccpay/requirements.txt
Django==2.2.16
pycrypto==2.6.1
pytz==2020.1
sqlparse==0.3.1
